libfock: cast memset sizes to size_t in LinK and DirectJK to avoid int overflow
The K-builder scratch-buffer memsets in LinK::build_G_component and
DirectJK::build_JK passed `nPbasis * nbf * sizeof(double)` and similar
products to memset. The `int * int` subexpression was evaluated in
int (the implicit widening to size_t happens only when sizeof(double)
joins in), so for systems where the product exceeds INT_MAX (~2.1e9 --
which here means nbf > ~50k with reasonable atom counts) the byte count
silently wrapped negative and was zero-extended to a garbage size_t.
The memset then either cleared far too few bytes (leaving stale entries
in K) or too many (potential overrun), depending on the wrap value.

Both call sites are in inner loops over shell pairs/quads, so the
intermediate dimensions can legitimately scale with system size. Fix by
casting the first operand to size_t so the whole product evaluates in
size_t arithmetic.

LinK is explicitly targeted at huge molecules where this matters most.
Current users with smaller systems are unaffected (int doesn't overflow).

+++ psi4/src/psi4/libfock/LinK.cc
@@ -380,10 +380,10 @@ void LinK::build_G_component(std::vector<std::shared_ptr<Matrix>>& D, std::vecto
                         const double* buffer2 = buffer;
 
                         if (!touched) {
-                            ::memset((void*)KTp[0L * max_functions_per_atom], '\0', nPbasis * nbf * sizeof(double));
-                            ::memset((void*)KTp[1L * max_functions_per_atom], '\0', nPbasis * nbf * sizeof(double));
-                            ::memset((void*)KTp[2L * max_functions_per_atom], '\0', nQbasis * nbf * sizeof(double));
-                            ::memset((void*)KTp[3L * max_functions_per_atom], '\0', nQbasis * nbf * sizeof(double));
+                            ::memset((void*)KTp[0L * max_functions_per_atom], '\0', static_cast<size_t>(nPbasis) * nbf * sizeof(double));
+                            ::memset((void*)KTp[1L * max_functions_per_atom], '\0', static_cast<size_t>(nPbasis) * nbf * sizeof(double));
+                            ::memset((void*)KTp[2L * max_functions_per_atom], '\0', static_cast<size_t>(nQbasis) * nbf * sizeof(double));
+                            ::memset((void*)KTp[3L * max_functions_per_atom], '\0', static_cast<size_t>(nQbasis) * nbf * sizeof(double));
                         }
 
                         // Four pointers needed for PR, PS, QR, QS
